refactor: Move LocalModCard into own Vue component
This is in preparation of overhauling the local mods interface
